requirements.txt
pymongo
python-dotenv
cryptography
tkinter